requirements.txt
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0	8.4.0
SQLAlchemy==1.4.27
WTForms==3.0.0
Werkzeug==2.0.2
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
lxml==4.6.4
pip==21.3.1
requests==2.26.0
setuptools==59.1.1
soupsieve==2.3.1
urllib3==1.26.7
wheel==0.37.0
